REF move some stuff to a new os_utils submodule
