+++ conda_forge_tick/migrators/conda_forge_yaml_cleanup.py
@@ -3,7 +3,7 @@ import typing
 from typing import Any
 from ruamel.yaml import YAML
 
-from conda_forge_tick.utils import pushd
+from conda_forge_tick.os_utils import pushd
 from conda_forge_tick.migrators.core import MiniMigrator
 
 if typing.TYPE_CHECKING:
