+++ conda_forge_tick/migrators/duplicate_lines.py
@@ -2,7 +2,7 @@ import typing
 from typing import Any
 import re
 
-from conda_forge_tick.utils import pushd
+from conda_forge_tick.os_utils import pushd
 from conda_forge_tick.migrators.core import MiniMigrator
 
 if typing.TYPE_CHECKING:
