+++ conda_forge_tick/migrators/mpi_pin_run_as_build.py
@@ -1,7 +1,7 @@
 import os
 import subprocess
 
-from conda_forge_tick.utils import pushd
+from conda_forge_tick.os_utils import pushd
 from conda_forge_tick.migrators.core import MiniMigrator
 
 MPIS = ["mpich", "openmpi"]
