+++ conda_forge_tick/os_utils.py
@@ -1,0 +1,38 @@
+import contextlib
+import copy
+import os
+import subprocess
+
+
+# https://stackoverflow.com/questions/6194499/pushd-through-os-system
+@contextlib.contextmanager
+def pushd(new_dir):
+    previous_dir = os.getcwd()
+    os.chdir(new_dir)
+    try:
+        yield
+    finally:
+        os.chdir(previous_dir)
+
+
+def eval_cmd(cmd, **kwargs):
+    """run a command capturing stdout
+
+    stderr is printed for debugging
+    any kwargs are added to the env
+    """
+    env = copy.deepcopy(os.environ)
+    timeout = kwargs.pop("timeout", None)
+    env.update(kwargs)
+    c = subprocess.run(
+        cmd,
+        shell=True,
+        stdout=subprocess.PIPE,
+        env=env,
+        timeout=timeout,
+    )
+    if c.returncode != 0:
+        print(c.stdout.decode("utf-8"), flush=True)
+        c.check_returncode()
+
+    return c.stdout.decode("utf-8")
